Revert "Made port numbers use portno_t."
This reverts commit 206745d4de20363c13574930b8bf9248d06710f0.

Conflicts:

	src/clustering/administration/main/command_line.cc
	src/http/json/json_adapter.cc
	src/http/json/json_adapter.hpp
	src/unittest/test_cluster_group.cc

+++ src/clustering/administration/main/ports.hpp
@@ -1,19 +1,17 @@
-// Copyright 2010-2013 RethinkDB, all rights reserved.
+// Copyright 2010-2012 RethinkDB, all rights reserved.
 #ifndef CLUSTERING_ADMINISTRATION_MAIN_PORTS_HPP_
 #define CLUSTERING_ADMINISTRATION_MAIN_PORTS_HPP_
 
-#include <stdint.h>
-
 namespace port_defaults {
-const uint16_t peer_port = 29015;
-const uint16_t client_port = 0;
-const uint16_t http_port = 8080;
+const int peer_port = 29015;
+const int client_port = 0;
+const int http_port = 8080;
 
 /* We currently spin up clusters with port offsets differing by 1, so
    these ports should be reasonably far apart. */
-const uint16_t reql_port = 28015;
+const int reql_port = 28015;
 
-const uint16_t port_offset = 0;
+const int port_offset = 0;
 }
 
 #endif  // CLUSTERING_ADMINISTRATION_MAIN_PORTS_HPP_
